refactor files to component 'microservices'
