+++ components/serverErrors.js
@@ -1,0 +1,5 @@
+class FailedToStartServerError extends Error {}
+
+module.exports = {
+  FailedToStartServerError,
+};
