+++ components/serverService.js
@@ -1,14 +1,9 @@
-const AWS = require('aws-sdk');
-const EC2Helper = require('../helpers/ec2Helper');
-const FailedToStartServerError = require('../models/errors/failedToStartServerError');
-
-const EC2Client = new AWS.EC2({
-  region: process.env.EC2_REGION,
-});
+const EC2 = require('../libraries/ec2');
+const FailedToStartServerError = require('./serverErrors');
 
 const getServerInformation = async (state, serverId) => {
-  const ec2Helper = new EC2Helper(EC2Client);
-  const serverInformation = await ec2Helper.describeInstance(state, serverId);
+  const ec2 = new EC2();
+  const serverInformation = await ec2.describeInstance(state, serverId);
 
   return serverInformation;
 };
@@ -17,8 +12,8 @@ const startServer = async (serverId) => {
   let { state, ipAddress } = await getServerInformation('instanceExists', serverId);
 
   if ((state === 'stopped') || (state === 'shutting-down')) {
-    const ec2Helper = new EC2Helper(EC2Client);
-    await ec2Helper.startInstance(serverId);
+    const ec2 = new EC2();
+    await ec2.startInstance(serverId);
 
     const serverInformation = await getServerInformation('instanceRunning', serverId);
     state = 'pending';
