--- models/errors/failedToStartServerError.js
@@ -1,1 +1,0 @@
-module.exports = class FailedToStartServerError extends Error {};
